add mock server for eip1559
